Add support for multivalue fields
This commit adds support for multivalue fields. I copied some code
wholesale from file_field_widget_submit to set field_state['items']
which we use in the widget form to populate the widget. Additionally, I
had to swap out the fid text field for a "refresh" button so that we could
take advantage of a submit callback for that button. I converted the fid
field to a hidden element and hid the button. The button is then
triggered to submit when the file browser dialog is closed, it submits
the form and the widget is rebuilt.