Add tabs styling and fixup cancel button Issue #70581
It turns out that there is a bunch of code that messes with primary tabs
and when you click cancel and then re-open the dialog the tabs disappear
because of that javascript. So I've copied the css for the tabs from the
seven theme. I also fixed the close event because if you just cancel the
file browser window, it shouldn't refresh the videos widget, otherwise
it will clear out all the videos because the fid is null.